# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * SakhiPath API - AI-powered career empowerment platform for women in STEM
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AiResponse {
  content: string;
  /** @nullable */
  sessionId?: string | null;
  /** @nullable */
  tokensUsed?: number | null;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface CareerAdviceInput {
  message: string;
  /** @nullable */
  sessionId?: string | null;
  /** @nullable */
  userId?: string | null;
  /** @nullable */
  careerStage?: string | null;
  /** @nullable */
  persona?: string | null;
  history?: ChatMessage[];
}

export interface ResumeFeedbackInput {
  resumeText: string;
  /** @nullable */
  targetRole?: string | null;
  /** @nullable */
  userId?: string | null;
}

export interface ResumeAnalysis {
  score: number;
  strengths: string[];
  weaknesses: string[];
  improvements: string[];
  summary: string;
}

export interface InterviewQuestionsInput {
  role: string;
  /** @nullable */
  level?: string | null;
  /** @nullable */
  userId?: string | null;
}

export interface LearningPlanInput {
  goals: string[];
  currentSkills: string[];
  /** @nullable */
  timeframe?: string | null;
  /** @nullable */
  userId?: string | null;
}

export interface ScholarshipRecoInput {
  profile: string;
  /** @nullable */
  fieldOfStudy?: string | null;
  /** @nullable */
  careerStage?: string | null;
  /** @nullable */
  userId?: string | null;
}

export interface FinancialLiteracyInput {
  topic: string;
  /** @nullable */
  level?: string | null;
  /** @nullable */
  userId?: string | null;
}

export interface CareerConfidenceInput {
  userId: string;
  completedModules: number;
  quizzesPassed: number;
  mentorSessions: number;
  aiSessions: number;
  profileCompleteness: number;
}

export type CareerConfidenceResultBreakdown = {
  learning: number;
  mentoring: number;
  aiEngagement: number;
  profileCompleteness: number;
};

export interface CareerConfidenceResult {
  score: number;
  level: string;
  insights: string;
  breakdown: CareerConfidenceResultBreakdown;
}

export type ConversationInputType = typeof ConversationInputType[keyof typeof ConversationInputType];


export const ConversationInputType = {
  career: 'career',
  resume: 'resume',
  interview: 'interview',
  learning: 'learning',
  financial: 'financial',
} as const;

export interface ConversationInput {
  userId: string;
  /** @nullable */
  title?: string | null;
  messages: ChatMessage[];
  type: ConversationInputType;
}

export interface Conversation {
  id: string;
  userId: string;
  /** @nullable */
  title?: string | null;
  messages: ChatMessage[];
  type: string;
  createdAt: string;
  updatedAt: string;
}

export interface Scholarship {
  id: number;
  title: string;
  provider: string;
  amount: string;
  /** @nullable */
  deadline?: string | null;
  eligibility: string;
  category: string;
  /** @nullable */
  applicationUrl?: string | null;
  description: string;
  isActive: boolean;
}

export interface ScholarshipList {
  items: Scholarship[];
  total: number;
}

export type LearningModuleDifficulty = typeof LearningModuleDifficulty[keyof typeof LearningModuleDifficulty];


export const LearningModuleDifficulty = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export interface QuizQuestion {
  question: string;
  options: string[];
  correctIndex: number;
  explanation: string;
}

export interface LearningModule {
  id: number;
  title: string;
  description: string;
  category: string;
  durationMinutes: number;
  difficulty: LearningModuleDifficulty;
  tags: string[];
  /** @nullable */
  content?: string | null;
  quizQuestions: QuizQuestion[];
}

export interface LearningModuleList {
  items: LearningModule[];
  total: number;
}

export interface QuizResultInput {
  userId: string;
  moduleId: number;
  score: number;
  totalQuestions: number;
  timeTakenSeconds?: number;
}

export interface QuizResult {
  id: number;
  userId: string;
  moduleId: number;
  score: number;
  totalQuestions: number;
  percentage: number;
  passed: boolean;
  createdAt: string;
}

export interface LearningProgress {
  userId: string;
  completedModules: number;
  totalModules: number;
  averageScore: number;
  badges: string[];
  recentResults: QuizResult[];
}

export type MentorAvailability = typeof MentorAvailability[keyof typeof MentorAvailability];


export const MentorAvailability = {
  available: 'available',
  limited: 'limited',
  unavailable: 'unavailable',
} as const;

export interface Mentor {
  id: number;
  name: string;
  title: string;
  organization: string;
  specializations: string[];
  bio: string;
  yearsExperience: number;
  rating: number;
  totalSessions: number;
  availability: MentorAvailability;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  photoUrl?: string | null;
}

export interface MentorList {
  items: Mentor[];
  total: number;
}

export interface MentorRequestInput {
  userId: string;
  mentorId: number;
  message: string;
  /** @nullable */
  preferredDate?: string | null;
}

export type MentorRequestStatus = typeof MentorRequestStatus[keyof typeof MentorRequestStatus];


export const MentorRequestStatus = {
  pending: 'pending',
  accepted: 'accepted',
  declined: 'declined',
  completed: 'completed',
} as const;

export interface MentorRequest {
  id: number;
  userId: string;
  mentorId: number;
  message: string;
  /** @nullable */
  preferredDate?: string | null;
  status: MentorRequestStatus;
  createdAt: string;
}

export interface MentorRequestList {
  items: MentorRequest[];
  total: number;
}

export type UpcomingDeadlineType = typeof UpcomingDeadlineType[keyof typeof UpcomingDeadlineType];


export const UpcomingDeadlineType = {
  scholarship: 'scholarship',
  learning: 'learning',
  mentor: 'mentor',
} as const;

export interface UpcomingDeadline {
  title: string;
  deadline: string;
  type: UpcomingDeadlineType;
}

export interface DashboardSummary {
  userId: string;
  careerConfidenceScore: number;
  completedModules: number;
  totalModules: number;
  activeMentorRequests: number;
  aiSessionsCount: number;
  scholarshipsBookmarked: number;
  upcomingDeadlines: UpcomingDeadline[];
  weeklyGoalProgress: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  ai_chat: 'ai_chat',
  quiz_completed: 'quiz_completed',
  mentor_request: 'mentor_request',
  scholarship_bookmark: 'scholarship_bookmark',
  module_completed: 'module_completed',
} as const;

export type ActivityItemMetadata = {
  /** @nullable */
  score?: number | null;
  /** @nullable */
  moduleName?: string | null;
};

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  description: string;
  timestamp: string;
  metadata: ActivityItemMetadata;
}

export interface ActivityFeed {
  items: ActivityItem[];
  total: number;
}

export type ListScholarshipsParams = {
category?: string;
search?: string;
limit?: number;
offset?: number;
};

export type ListLearningModulesParams = {
category?: string;
};

export type ListMentorsParams = {
specialization?: string;
search?: string;
};