# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SakhiPath API - AI-powered career empowerment platform for women in STEM
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scholarships:
    get:
      operationId: listScholarships
      tags: [scholarships]
      summary: List scholarships with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of scholarships
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScholarshipList"
  /scholarships/{id}:
    get:
      operationId: getScholarship
      tags: [scholarships]
      summary: Get scholarship details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scholarship details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scholarship"

  # ── Learning ─────────────────────────────────────────────────────────────────
...
  /mentors:
    get:
      operationId: listMentors
      tags: [mentors]
      summary: List mentors with optional filters
      parameters:
        - name: specialization
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of mentors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MentorList"
  /mentors/requests:
    post:
      operationId: createMentorRequest
      tags: [mentors]
      summary: Request a mentor session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MentorRequestInput"
      responses:
        "201":
          description: Mentor request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MentorRequest"
  /mentors/requests/{userId}:
    get:
      operationId: getMentorRequests
      tags: [mentors]
      summary: Get mentor requests for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of mentor requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MentorRequestList"

  # ── Dashboard ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Scholarship:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        provider:
          type: string
        amount:
          type: string
        deadline:
          type: ["string", "null"]
        eligibility:
          type: string
        category:
          type: string
        applicationUrl:
          type: ["string", "null"]
        description:
          type: string
        isActive:
          type: boolean
      required:
        - id
        - title
        - provider
        - amount
        - eligibility
        - category
        - description
        - isActive
    ScholarshipList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Scholarship"
        total:
          type: integer
      required:
        - items
        - total

    # ── Learning schemas ────────────────────────────────────────────────────────
...
    Mentor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        title:
          type: string
        organization:
          type: string
        specializations:
          type: array
          items:
            type: string
        bio:
          type: string
        yearsExperience:
          type: integer
        rating:
          type: number
        totalSessions:
          type: integer
        availability:
          type: string
          enum: [available, limited, unavailable]
        linkedinUrl:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
      required:
        - id
        - name
        - title
        - organization
        - specializations
        - bio
        - yearsExperience
        - rating
        - totalSessions
        - availability
    MentorList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Mentor"
        total:
          type: integer
      required:
        - items
        - total
    MentorRequestInput:
      type: object
      properties:
        userId:
          type: string
        mentorId:
          type: integer
        message:
          type: string
        preferredDate:
          type: ["string", "null"]
      required:
        - userId
        - mentorId
        - message
    MentorRequest:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        mentorId:
          type: integer
        message:
          type: string
        preferredDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, accepted, declined, completed]
        createdAt:
          type: string
      required:
        - id
        - userId
        - mentorId
        - message
        - status
        - createdAt
    MentorRequestList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MentorRequest"
        total:
          type: integer
      required:
        - items
        - total

    # ── Dashboard schemas ───────────────────────────────────────────────────────
...