# lib/api-zod/src/generated/types/mentorRequestInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * SakhiPath API - AI-powered career empowerment platform for women in STEM
 * OpenAPI spec version: 0.1.0
 */

export interface MentorRequestInput {
  userId: string;
  mentorId: number;
  message: string;
  /** @nullable */
  preferredDate?: string | null;
}