# lib/api-zod/src/generated/types/scholarshipList.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * SakhiPath API - AI-powered career empowerment platform for women in STEM
 * OpenAPI spec version: 0.1.0
 */
import type { Scholarship } from './scholarship';

export interface ScholarshipList {
  items: Scholarship[];
  total: number;
}